requirements.txt
Flask
flask-cors
scikit-learn
joblib
pandas
numpy
pyjwt
